English Dictionary/requirements.txt
pyenchant~=3.2.2
